requirements.txt
Flask==1.1.2
Flask-Assets==2.0
google-api-python-client==2.55.0
Hamlish-Jinja==0.3.3
Jinja2==2.11.2
python-dotenv==0.20.0
